chore(sb/preview): split decorators into own file
